Extract kanji_vocab_algo reporting into kanji_vocab_report
(c) from the review — move the ~280 lines of stdout reporting (tier/length/JLPT
breakdowns, reading-diversity stats, the emoji column-width padding helpers, the
replacement-log table) out of kanji_vocab_algo into a sibling kanji_vocab_report
module, so the selection algorithm reads top-to-bottom without formatting noise.

The report is diagnostic-only — it prints to stdout and writes no file — so this
is pure relocation with no effect on shipped output. Selection state it needs
(SHIPPED, WORD_JLPT, TAG_PRIORITY, kanji_reading_in_word, has_nonshipped_kanji)
stays on kanji_vocab_algo and is read through a `kv` alias; the report is imported
late inside main() so the module's top-level `import kanji_vocab_algo` sees a fully
loaded module (no import cycle).

kanji_vocab_algo.py: 996 -> 719 lines. Also drops the now-unused unicodedata import.

Verified: byte-compile, no dangling references, and offline harnesses that run
both print_report + print_replace_logs (all sections render; kv.SHIPPED /
kv.WORD_JLPT read correctly) and the selection path (unchanged).
